Refactorización de los tests de comandos: ahora se controlan mejor ambos escenarios (fichero nuevo y fichero existente), y se limpian los ficheros troquelados.
