+++ dgt-envlabel-api/api/.env.test
@@ -6,3 +6,12 @@ PANTHER_APP_ENV=panther
 PANTHER_ERROR_SCREENSHOT_DIR=./var/error-screenshots
 
 DATABASE_URL="sqlite:///%kernel.project_dir%/var/test.db"
+
+###> App stuff ###
+TMP_STORAGE_DIR='/tmp/'
+DGT_ENVIRONMENTAL_LABELS_URL='https://sedeapl.dgt.gob.es/IEST_INTER/MICRODATOS/salida/distintivoAmbiental/'
+DGT_ENVIRONMENTAL_LABELS_ZIP_FILE='export_dist_ambiental.zip'
+DGT_ENVIRONMENTAL_LABELS_CSV_FILE='export_distintivo_ambiental.txt'
+SPLIT_COMMAND_LINES_ARG=20000
+SPLIT_COMMAND_STORAGE_DIR='/tmp/splits/'
+###< App stuff ###
