+++ dgt-envlabel-api/api/tests/Command/DownloadNewFileCommandTest.php
@@ -4,12 +4,18 @@ declare(strict_types=1);
 
 namespace App\Tests\Command;
 
+use Hautelook\AliceBundle\PhpUnit\RecreateDatabaseTrait;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Component\Console\Tester\CommandTester;
+use Symfony\Component\Dotenv\Dotenv;
+use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\Finder\Finder;
+use Symfony\Component\Messenger\Transport\InMemoryTransport;
 
-class DownloadDgtEnvLabelsFileCommandTest extends KernelTestCase
+class DownloadNewFileCommandTest extends KernelTestCase
 {
+    use RecreateDatabaseTrait;
 
     public function testDownloadDgtEnvLabelsFileNewFile_ok(): void
     {
@@ -22,6 +28,10 @@ class DownloadDgtEnvLabelsFileCommandTest extends KernelTestCase
 
         $commandTester->assertCommandIsSuccessful();
 
+        /** @var InMemoryTransport $transport */
+        $transport = self::getContainer()->get('messenger.transport.amqp_csv');
+        $this->assertGreaterThanOrEqual(1, $transport->get());
+
         $output = $commandTester->getDisplay();
         $this->assertStringContainsString('let\'s unzip & split the downloaded file!', $output);
         $this->assertStringContainsString('let\'s send those messages to RabbitMQ!', $output);
@@ -29,23 +39,23 @@ class DownloadDgtEnvLabelsFileCommandTest extends KernelTestCase
         $this->assertStringContainsString('Downloaded file hash has been logged!', $output);
         $this->assertStringContainsString('Downloaded zip file has been deleted successfully.', $output);
         $this->assertStringContainsString('app:download-environmental-labels-file → Finished', $output);
+
+        $this->cleanSplitFilesIfNecessary();
     }
 
-    public function testDownloadDgtEnvLabelsFileAlreadyProcessedFile_ok(): void
+    private function cleanSplitFilesIfNecessary(): void
     {
-        $kernel = self::bootKernel();
-        $application = new Application($kernel);
+        $filesystem = new Filesystem();
+        $finder = new Finder();
 
-        $command = $application->find('app:download-environmental-labels-file');
-        $commandTester = new CommandTester($command);
-        $commandTester->execute([]);
+        $finder->files()->in($_ENV['SPLIT_COMMAND_STORAGE_DIR'])
+            ->name([sprintf('%s-*', $_ENV['DGT_ENVIRONMENTAL_LABELS_CSV_FILE'])]);
 
-        $commandTester->assertCommandIsSuccessful();
-
-        $output = $commandTester->getDisplay();
-        $this->assertStringContainsString('Downloaded file is identical to the previous one', $output);
-        $this->assertStringContainsString('Downloaded zip file has been deleted successfully.', $output);
-        $this->assertStringContainsString('app:download-environmental-labels-file → Finished', $output);
+        foreach ($finder as $file) {
+            if ($filesystem->exists($file->getRealPath())) {
+                $filesystem->remove($file);
+            }
+        }
     }
 
 }
